+++ src/test/Fake.Core.IntegrationTests/TestHelpers.fs
@@ -27,11 +27,13 @@ let prepare scenario =
 
 let directFakeInPath command scenarioPath target =
     let result =
-        Process.ExecProcessAndReturnMessages (fun (info:System.Diagnostics.ProcessStartInfo) ->
-          info.EnvironmentVariables.["target"] <- target
-          info.FileName <- fakeToolPath
-          info.WorkingDirectory <- scenarioPath
-          info.Arguments <- command) (System.TimeSpan.FromMinutes 15.)
+        Process.ExecProcessAndReturnMessages (fun (info:Process.ProcStartInfo) ->
+          { info with
+                FileName = fakeToolPath
+                WorkingDirectory = scenarioPath
+                Arguments = command }
+          |> Process.setEnvironmentVariable "target" target
+          |> Process.setEnvironmentVariable "FAKE_DETAILED_ERRORS" "true") (System.TimeSpan.FromMinutes 15.)
     if result.ExitCode <> 0 then 
         let errors = String.Join(Environment.NewLine,result.Errors)
         printfn "%s" <| String.Join(Environment.NewLine,result.Messages)